+++ launchpy/__init__.py
@@ -9,3 +9,4 @@ from .configs import *
 from .library import *
 from .admin import *
 from .property import *
+from .synchronizer import *